requirements.txt
pandas==2.2.3
plotly==5.24.1
python_dateutil==2.9.0.post0
streamlit==1.39.0